cleanup directories
