+++ Bot/cogs/add.py
