+++ Bot/cogs/delete.py
