+++ Bot/cogs/search.py
