--- Bot/creds/README.md
@@ -1,7 +1,0 @@
-# This Folder contains the Google Drive API credentials
-
-Two files:
-a) credentials.json
-b) token.json
-
-For more info check: <a href=https://github.com/Servatom/Notion-DiscordBot/tree/main/GoogleDrive_Setup>Google Drive Setup</a>
