+++ Bot/functionality/deleteRecord.py
