+++ Bot/functionality/search.py
