+++ Bot/functionality/setupBot.py
