+++ Bot/functionality/utils.py
