--- botV2/requirements.txt
@@ -1,46 +1,0 @@
-aiohttp==3.7.4.post0
-async-timeout==3.0.1
-attrs==21.2.0
-beautifulsoup4==4.9.3
-bs4==0.0.1
-cachetools==4.2.2
-certifi==2021.5.30
-chardet==4.0.0
-decorator==5.0.9
-discord==1.7.3
-discord.py==1.7.3
-fuzzywuzzy==0.18.0
-google-api-core==1.30.0
-google-api-python-client==2.12.0
-google-auth==1.32.1
-google-auth-httplib2==0.1.0
-google-auth-oauthlib==0.4.4
-googleapis-common-protos==1.53.0
-httplib2==0.19.1
-idna==2.10
-multidict==5.1.0
-oauth2client==4.1.3
-oauthlib==3.1.1
-packaging==21.0
-ply==3.11
-protobuf==3.17.3
-pyasn1==0.4.8
-pyasn1-modules==0.2.8
-PyDrive==1.3.1
-pyparsing==2.4.7
-python-Levenshtein==0.12.2
-python-magic==0.4.24
-pytz==2021.1
-PyYAML==5.4.1
-requests==2.25.1
-requests-oauthlib==1.3.0
-rsa==4.7.2
-six==1.16.0
-soupsieve==2.2.1
-SQLAlchemy==1.4.28
-stone==3.2.1
-typing-extensions==3.10.0.0
-uritemplate==3.0.1
-urllib3==1.26.6
-validators==0.18.2
-yarl==1.6.3
