+++ docker-compose.yml
@@ -8,10 +8,7 @@ services:
       context: .
       dockerfile: ./Dockerfile
     volumes:
-      - /home/user/directory_where_creds_kept:/Bot/creds
+      - ./database:/Bot/database
     environment: 
-      - DATABASE_TOKEN=your-notion-database-token
-      - AUTH_KEY=your-notion-integration-key
-      - DISCORD_AUTH=your-discord-bot-token
-      - GDRIVE_FOLDER=<FOLDER_ID>
-      - PREFIX=/
+      - TOKEN=your-discord-bot-token
+      - PREFIX=*
