+++ v1/Bot/.env.example
