+++ v1/Bot/delete.py
